pyproject.toml
[project]
name = "ai-mentor-hub"
version = "0.1.0"
requires-python = "==3.12.*"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.3",
    "aiosignal==1.4.0",
    "altair==6.0.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "attrs==25.4.0",
    "blinker==1.9.0",
    "cachetools==7.0.5",
    "certifi==2026.2.25",
    "cffi==2.0.0",
    "charset-normalizer==3.4.5",
    "click==8.3.1",
    "cryptography==46.0.5",
    "dataclasses-json==0.6.7",
    "ddgs>=9.11.4",
    "distro==1.9.0",
    "docarray==0.41.0",
    "duckduckgo-search>=8.1.1",
    "filelock==3.25.2",
    "filetype==1.2.0",
    "frozenlist==1.8.0",
    "fsspec==2026.2.0",
    "gitdb==4.0.12",
    "gitpython==3.1.46",
    "google-auth==2.49.1",
    "google-genai==1.67.0",
    "greenlet==3.3.2",
    "h11==0.16.0",
    "hf-xet==1.4.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "huggingface-hub==0.36.2",
    "idna==3.11",
    "iniconfig==2.3.0",
    "jinja2==3.1.6",
    "joblib==1.5.3",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.26.0",
    "jsonschema-specifications==2025.9.1",
    "langchain==1.2.12",
    "langchain-classic==1.0.2",
    "langchain-community==0.4.1",
    "langchain-core==1.2.18",
    "langchain-google-genai==4.2.1",
    "langchain-huggingface==1.2.1",
    "langchain-text-splitters==1.1.1",
    "langchainhub>=0.1.18",
    "langdetect>=1.0.9",
    "langgraph==1.1.2",
    "langgraph-checkpoint==4.0.1",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "langgraph-prebuilt==1.0.8",
    "langgraph-sdk==0.3.11",
    "langsmith==0.7.17",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "marshmallow==3.26.2",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "multidict==6.7.1",
    "mypy-extensions==1.1.0",
    "narwhals==2.18.0",
    "networkx==3.6.1",
    "numpy==1.26.4",
    "orjson==3.11.7",
    "ormsgpack==1.12.2",
    "packaging==26.0",
    "pandas==2.3.3",
    "pillow==12.1.1",
    "pluggy==1.6.0",
    "propcache==0.4.1",
    "protobuf==6.33.5",
    "pyarrow==14.0.2",
    "pyasn1==0.6.2",
    "pyasn1-modules==0.4.2",
    "pycparser==3.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.13.1",
    "pydeck==0.9.1",
    "pygments==2.19.2",
    "pypdf==6.8.0",
    "pytest==9.0.2",
    "pytest-mock==3.15.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.2",
    "pytz==2026.1.post1",
    "pyyaml==6.0.3",
    "referencing==0.37.0",
    "regex==2026.2.28",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "rich==14.3.3",
    "rpds-py==0.30.0",
    "safetensors==0.7.0",
    "scikit-learn==1.8.0",
    "scipy==1.17.1",
    "sentence-transformers==2.7.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smmap==5.0.3",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.48",
    "streamlit==1.55.0",
    "sympy==1.14.0",
    "tenacity==9.1.4",
    "threadpoolctl==3.6.0",
    "tokenizers==0.19.1",
    "toml==0.10.2",
    "torch==2.2.2",
    "tornado==6.5.5",
    "tqdm==4.67.3",
    "transformers==4.44.2",
    "typer==0.24.1",
    "types-requests==2.32.4.20260107",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.3",
    "urllib3==2.6.3",
    "uuid-utils==0.14.1",
    "websockets==16.0",
    "xxhash==3.6.0",
    "yarl==1.23.0",
    "zstandard==0.25.0",
]
